+++ client/graph.c
@@ -14,6 +14,7 @@
 #include "ui.h"
 #include "graph.h"
 #include "lfdemod.h"
+#include "cmddata.h" //for g_debugmode
 
 int GraphBuffer[MAX_GRAPH_TRACE_LEN];
 int GraphTraceLen;
@@ -255,7 +256,7 @@ uint8_t fskClocks(uint8_t *fc1, uint8_t *fc2, uint8_t *rf1, bool verbose)
 	if (size==0) return 0;
 	uint16_t ans = countFC(BitStream, size, 1); 
 	if (ans==0) {
-		if (verbose) PrintAndLog("DEBUG: No data found");
+		if (verbose || g_debugMode) PrintAndLog("DEBUG: No data found");
 		return 0;
 	}
 	*fc1 = (ans >> 8) & 0xFF;
@@ -263,7 +264,7 @@ uint8_t fskClocks(uint8_t *fc1, uint8_t *fc2, uint8_t *rf1, bool verbose)
 
 	*rf1 = detectFSKClk(BitStream, size, *fc1, *fc2);
 	if (*rf1==0) {
-		if (verbose) PrintAndLog("DEBUG: Clock detect error");
+		if (verbose || g_debugMode) PrintAndLog("DEBUG: Clock detect error");
 		return 0;
 	}
 	return 1;